assignment number 2
implement functions for caching the computations of the
inverse of a matrix
the first function is returning an object that caches a list
of four functions and the initial matrix and the inverse
(if computed)

the computation is actually done in the second function, if
it was not cached already. It will also save the state.